Add requirements analysis TeX and PDF
With this commit, the initial requirements doc is added with the basic
skeleton added.  Most of the first chapters are already populated with
initial requirements, but are by no means complete. A .gitignore file is
also added to filter out unnecessary documents that are created as a
result of the TeX compilation process.